+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.0.2] - 2021-10-01
+
+### Fixes
+- Moved axios to dev dependency
+- Fixed axios refresh error test
+
+### Changes
+- Using fetch instead of axios to call the refresh API
+
 ## [9.0.1] - 2021-10-01
 
 ### Fixes